Updated step file
