--- src/test/java/com/znsio/teswiz/steps/VodQaSteps.java
@@ -1,54 +1,0 @@
-package com.znsio.teswiz.steps;
-
-import com.context.SessionContext;
-import com.context.TestExecutionContext;
-import com.znsio.teswiz.businessLayer.vodqa.VodqaBL;
-import com.znsio.teswiz.entities.SAMPLE_TEST_CONTEXT;
-import com.znsio.teswiz.runner.Drivers;
-import com.znsio.teswiz.runner.Runner;
-import io.cucumber.java.en.And;
-import io.cucumber.java.en.Given;
-import io.cucumber.java.en.Then;
-import io.cucumber.java.en.When;
-import org.apache.log4j.Logger;
-
-public class VodQASteps {
-    private static final Logger LOGGER = Logger.getLogger(VodQASteps.class.getName());
-    private final TestExecutionContext context;
-
-    public VodQASteps() {
-        context = SessionContext.getTestExecutionContext(Thread.currentThread().getId());
-        LOGGER.info("context: " + context.getTestName());
-    }
-
-    @Given("I login to vodqa application using valid credentials")
-    public void loginToApplication() {
-        Drivers.createDriverFor(SAMPLE_TEST_CONTEXT.ME, Runner.getPlatform(), context);
-        new VodqaBL(SAMPLE_TEST_CONTEXT.ME, Runner.getPlatform()).login();
-    }
-
-    @When("I scroll from one to another element point on vertical swiping screen")
-    public void scrollToElement() {
-        new VodqaBL(SAMPLE_TEST_CONTEXT.ME, Runner.getPlatform()).scrollFromOneElementPointToAnother();
-    }
-
-    @Then("Element text should be visible")
-    public void isElementWithTextVisible() {
-        new VodqaBL(SAMPLE_TEST_CONTEXT.ME, Runner.getPlatform()).isElementWithTextVisible();
-    }
-
-    @When("I enter into hacker news under webview section")
-    public void iEnterIntoHackerNewsUnderWebviewSection() {
-        new VodqaBL(SAMPLE_TEST_CONTEXT.ME, Runner.getPlatform()).enterIntoNewsWebViewSection();
-    }
-
-    @Then("I am able to see hacker news login option in webview section")
-    public void iAmAbleToSeeHackerNewsLoginOptionInWebViewSection() {
-        new VodqaBL(SAMPLE_TEST_CONTEXT.ME, Runner.getPlatform()).verifyLoginOptionForHackerNewsWebview();
-    }
-
-    @And("I am able to navigate back to native view section")
-    public void iAmAbleToNavigateBackToNativeViewSection() {
-        new VodqaBL(SAMPLE_TEST_CONTEXT.ME, Runner.getPlatform()).enterIntoNativeViewSection();
-    }
-}
